Fully Fixed (pull before you start working), read description
This commit has the animated player sprite, the art for the witch hut roof and the code works properly. I also did some cleaning of the file structure. Lets try to keep things in their own folders. DO NOT CHANGE THE FOLDERS FOR SCENES AS THEIR LOCATIONS ARE MENTIONED IN SCRIPTS